requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.1.0
aiosqlite==0.22.1
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apache-airflow==3.0.6
apache-airflow-core==3.0.6
apache-airflow-providers-common-compat==1.12.0
apache-airflow-providers-common-io==1.7.1
apache-airflow-providers-common-sql==1.30.3
apache-airflow-providers-smtp==2.4.2
apache-airflow-providers-standard==1.10.3
apache-airflow-task-sdk==1.0.6
argcomplete==3.6.3
asgiref==3.11.0
attrs==25.4.0
blinker==1.9.0
build==1.4.0
CacheControl==0.14.4
cadwyn==6.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
crashtest==0.4.1
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.4
Deprecated==1.3.1
dill==0.4.1
distlib==0.4.0
dnspython==2.8.0
dulwich==0.25.2
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastjsonschema==2.21.2
filelock==3.20.3
findpython==0.7.1
Flask==3.1.2
fsspec==2026.1.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
gunicorn==24.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.0
keyring==25.7.0
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgpack==1.1.2
msgspec==0.20.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pathlib_abc==0.5.2
pathspec==1.0.4
pbs-installer==2026.1.27
pendulum==3.2.0
pkginfo==1.12.1.2
platformdirs==4.5.1
pluggy==1.6.0
poetry==2.3.1
poetry-core==2.3.0
protobuf==6.33.5
psutil==7.2.2
py4j==0.10.9.9
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
pyspark==4.1.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
retryhttp==1.4.0
rich==14.3.1
rich-argparse==1.7.2
rich-toolkit==0.17.2
rignore==0.7.6
rpds-py==0.30.0
sentry-sdk==2.51.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.50.0
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.3.0
text-unidecode==1.3
tomlkit==0.14.0
trove-classifiers==2026.1.14.14
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
universal_pathlib==0.3.8
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.40.0
virtualenv==20.36.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
wheel==0.46.3
wirerope==1.0.0
wrapt==2.0.1
zipp==3.23.0
zstandard==0.25.0
